requirements/base.txt
# Core
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.2

# ML & LLM
torch==2.3.0
transformers==4.41.2
accelerate==0.31.0
huggingface_hub==0.23.4
sentence-transformers==3.0.1
chromadb==0.4.24           # Более старая версия, совместимая с httpx 0.26
langchain==0.2.5
langchain-community==0.2.5

# Vision
opencv-python==4.9.0.80
mss==9.0.1
pillow==10.3.0

# Audio
pydub==0.25.1
sounddevice==0.4.6
openai-whisper @ git+https://github.com/openai/whisper.git

# Documents
PyPDF2==3.0.1
python-docx==1.1.2
openpyxl==3.1.2
python-pptx==0.6.23
markdown==3.6

# Video
moviepy==1.0.3

# Interfaces
python-telegram-bot==20.8
fastapi==0.111.0
uvicorn[standard]==0.30.1
selenium==4.21.0
webdriver-manager==4.0.1

# HTTP клиент (общий для всех)
httpx==0.26.0              # Фиксированная версия для совместимости

# Utils
watchdog==4.0.1
psutil==5.9.8